userspace-dp: extract ZoneScreenState + two-phase SYN-flood gate into screen/zone.rs (#6437)
check_packet_with_zone_id_opts was a ~315-line per-packet god function
with the ~155-line SYN-flood / SYN-cookie enforcement block still
inlined (#6238 extracted the stateless tail and the shared ICMP/UDP
rate-flood block; this is the residual). A per-packet function that
long silently degrades reviewability and inlining, so the screen
module gets the same split before the next change lands on it
(docs/engineering-style.md modularity discipline).

Pure code-motion within the crate, split at the borrow boundary:

- screen/zone.rs (new) holds ZoneScreenState (the #4969 consolidated
  per-zone value, moved verbatim — field order preserved), its
  from_profile / reconcile_substate constructors, the ICMP/UDP flood
  admission helpers, the #6238 enforce_common_rate_floods block,
  reconcile_flood_sketch, SECONDARY_FLOOD_CEILING_MULT, and the new
  ZoneScreenState::syn_flood_gate.
- syn_flood_gate is PHASE 1 of the enforcement: it performs every
  zone-local mutation in the exact historical #3315/#4112-F19 order
  (aggregate increment_and_classify → per-destination sketch →
  aggregate over-attack verdict incl. the cookie active-until stamp
  and the alarm_without_drop audit gate → alarm cadence → per-source
  sketch) and returns an owned SynFloodGate enum, so the &mut zones
  borrow ends at the call. check_packet_with_zone_id_opts then runs
  PHASE 2 — the whole-ScreenState applications (syn_alarm_pending,
  syn_flood_alarm_events, syn_flood_{dst,src}_drops, the cookie mint
  via current_syn_cookie_full_epoch) — with the borrow released. The
  validated-cache consume stays in the caller as PHASE 0 (it reads a
  disjoint self field while zstate is live).
- codec_available is a caller-side snapshot of
  self.syn_cookie_codec.is_some() so the gate picks MintChallenge vs
  DropCookieUnavailable without borrowing the codec; the caller
  re-checks the codec at the mint site, keeping the fail-closed
  "syn-cookie-unavailable" drop identical.

Preserved and pinned by the unmodified screen test suites: drop
precedence and reason strings (the screen_reason_drop_index ordinals
are untouched), every counter increment site, the count-min sketch
semantics (#3315), the cookie-OFF TokenBucket sole-drop-authority
regime (#3607), the alarm ≤1/sec/zone cadence, the audit-mode
contract (no active-until stamp, per-source still runs), and the
#3527 tcp_opening_ns override consumers (outside this function).

Hot-path shape: syn_flood_gate keeps a plain #[inline] hint (the
#6238 precedent); LLVM outlines it at the single callsite, so the
per-initial-SYN enforcement now sits behind one direct call — the
same posture the base build already had for
enforce_common_rate_floods. Asm-diff of
ScreenState::check_packet_with_zone_id_opts (release, base vs
patched): 478 → 457 instructions in the hot fn with the enforcement
block contiguous in the outlined gate; identical callee set
(take_valid → gate → current_syn_cookie_full_epoch → mint_isn); the
PHASE-2 match lowers to a jump table; reason-string immediates
(9 = "syn-flood", 22 = "syn-cookie-unavailable") unchanged; no new
allocation, lock, atomic, or dynamic dispatch on any path.

Validation: cargo test --release --test-threads=1 full suite GREEN
(4290 passed / 0 failed on origin/master 4bc33a3b0, screen suites
unmodified); cargo check --all-targets shows no new warnings vs
base. Docs: docs/syn-cookie-flood-protection.md enforcement-order
and #4969 citations repointed at screen/zone.rs;
userspace-dp/README.md's stale src/screen.rs row predates this
change (screen/ has been a directory since the tests relocation)
and is left alone.

Closes #6437
